fix(fscomponents): Move story files to all live in the same folder
Them being in different folders was making them load in unpredictable orders, so they were no longer alphabetized.
