+++ packages/fscomponents/src/components/__stories__/Form.story.tsx
@@ -10,9 +10,9 @@ import {
 } from '@storybook/addon-knobs/react';
 // @ts-ignore: TODO fix types for tcomb-form-native
 import * as t from 'tcomb-form-native';
-import { Form, FormLabelPosition } from '../Form';
-import { Button } from '../../Button';
-import { emailRegex } from '../../../lib/email';
+import { Form, FormLabelPosition } from '../Form/Form';
+import { Button } from '../Button';
+import { emailRegex } from '../../lib/email';
 
 const EmailType = t.refinement(t.String, (str: string) => {
   return emailRegex.test((str || '').trim());
