+++ packages/fscomponents/src/components/__stories__/MultiCarousel.story.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react'; // tslint:disable-line:no-implicit-dependencies
 import { action } from '@storybook/addon-actions'; // tslint:disable-line:no-implicit-dependencies
-import { MultiCarousel } from '../MultiCarousel';
-import { ProductItem } from '../../ProductItem';
+import { MultiCarousel } from '../MultiCarousel/MultiCarousel';
+import { ProductItem } from '../ProductItem';
 import Decimal from 'decimal.js';
 import {
   boolean
